OP-7470 - moved around only proper fields
Some fields (frame and frame_range) are making sense only in specific creator.
