+++ openpype/hosts/fusion/api/plugin.py
@@ -185,13 +185,7 @@ class GenericCreateSaver(Creator):
 
     def get_pre_create_attr_defs(self):
         """Settings for create page"""
-        attr_defs = [
-            self._get_render_target_enum(),
-            self._get_reviewable_bool(),
-            self._get_frame_range_enum(),
-            self._get_image_format_enum(),
-        ]
-        return attr_defs
+        raise NotImplementedError
 
     def get_instance_attr_defs(self):
         """Settings for publish page"""
@@ -202,8 +196,6 @@ class GenericCreateSaver(Creator):
         for pass_key in pre_create_data.keys():
             creator_attrs[pass_key] = pre_create_data[pass_key]
 
-    # These functions below should be moved to another file
-    # so it can be used by other plugins. plugin.py ?
     def _get_render_target_enum(self):
         rendering_targets = {
             "local": "Local machine rendering",
@@ -214,19 +206,6 @@ class GenericCreateSaver(Creator):
 
         return EnumDef(
             "render_target", items=rendering_targets, label="Render target"
-        )
-
-    def _get_frame_range_enum(self):
-        frame_range_options = {
-            "asset_db": "Current asset context",
-            "render_range": "From render in/out",
-            "comp_range": "From composition timeline",
-        }
-
-        return EnumDef(
-            "frame_range_source",
-            items=frame_range_options,
-            label="Frame range source",
         )
 
     def _get_reviewable_bool(self):
