+++ openpype/hosts/fusion/plugins/create/create_saver.py
@@ -1,3 +1,5 @@
+from openpype.lib import EnumDef
+
 from openpype.hosts.fusion.api.plugin import GenericCreateSaver
 
 
@@ -12,6 +14,8 @@ class CreateSaver(GenericCreateSaver):
     name = "render"
     family = "render"
     description = "Fusion Saver to generate image sequence"
+
+    default_frame_range_option = "asset_db"
 
     def get_detail_description(self):
         return """Fusion Saver to generate image sequence.
@@ -33,3 +37,27 @@ class CreateSaver(GenericCreateSaver):
         - tif
         - jpg
         """
+
+    def get_pre_create_attr_defs(self):
+        """Settings for create page"""
+        attr_defs = [
+            self._get_render_target_enum(),
+            self._get_reviewable_bool(),
+            self._get_frame_range_enum(),
+            self._get_image_format_enum(),
+        ]
+        return attr_defs
+
+    def _get_frame_range_enum(self):
+        frame_range_options = {
+            "asset_db": "Current asset context",
+            "render_range": "From render in/out",
+            "comp_range": "From composition timeline",
+        }
+
+        return EnumDef(
+            "frame_range_source",
+            items=frame_range_options,
+            label="Frame range source",
+            default=self.default_frame_range_option
+        )
